Python support for arbitrary sequences of hashable objects (#128)
This is a workaround until https://github.com/Martinsos/edlib/issues/90
is implemented.
If either query or target contain non-ascii values, they are mapped into
an ASCII alphabet and the resulting byte sequences are used for doing
the alignment. This works only if whole alphabet does not have more than 256 characters.