Updated the rest of the cpp code: tests, aligner to work with the changes made in PR #148.
Template parameteres are added to the tests for random sequences. Now they can be called with different Element types.
A specific test is added that checks if AlphabetTooBigException is thrown when it has to.
The aligner app is back to work with minimal changes, it supports just character sequences as before.