Remove some unnecessary allocations from transformSequences()
`alphabet += ...` isn't a trivial operation since it has to check if it
needs to allocate more memory each time it's called. Since the alphabet
can't include any duplicates we can instead create a fixed size buffer
and build it on the stack, then perform a single allocation at the end.

Also move the storing of the transformed pointers to the end so that
the compiler can infer that they don't alias with anything.

This gives another small boost in performance.