Drop some unnecessary allocations (#230)
* Drop an unnecessary allocation

Putting containers on the heap doesn't do anything since they allocate
internally and don't grow on the stack, so it's just additional work
having to call into the allocator.

* Replace a fixed size vector with an array

There's only 64 int's which is 256 bytes on most platforms.

This improves performance by a small but measureable amount in the
tests.

* Remove some unnecessary allocations from transformSequences()

`alphabet += ...` isn't a trivial operation since it has to check if it
needs to allocate more memory each time it's called. Since the alphabet
can't include any duplicates we can instead create a fixed size buffer
and build it on the stack, then perform a single allocation at the end.

Also move the storing of the transformed pointers to the end so that
the compiler can infer that they don't alias with anything.

This gives another small boost in performance.